+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/installerorder/OrderedInstallersApplication.groovy
@@ -4,7 +4,6 @@ import io.dropwizard.Application
 import io.dropwizard.setup.Bootstrap
 import io.dropwizard.setup.Environment
 import ru.vyarus.dropwizard.guice.GuiceBundle
-import ru.vyarus.dropwizard.guice.module.installer.WebInstallersBundle
 import ru.vyarus.dropwizard.guice.support.TestConfiguration
 import ru.vyarus.dropwizard.guice.support.feature.DummyResource
 import ru.vyarus.dropwizard.guice.support.feature.DummyService
@@ -16,13 +15,13 @@ import ru.vyarus.dropwizard.guice.support.util.BindModule
  * @author Vyacheslav Rusakov 
  * @since 17.04.2015
  */
-class OrderedInstallersApplication extends Application<TestConfiguration>{
+class OrderedInstallersApplication extends Application<TestConfiguration> {
 
     @Override
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
         bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
                 .enableAutoConfig("ru.vyarus.dropwizard.guice.support.installerorder")
-                .bundles(new WebInstallersBundle())
+                .useWebInstallers()
                 // at least one resource required
                 .extensions(DummyResource)
                 .modules(new BindModule(DummyService))