+++ src/test/groovy/ru/vyarus/dropwizard/guice/support/web/ServletsApplication.groovy
@@ -4,7 +4,6 @@ import io.dropwizard.Application
 import io.dropwizard.setup.Bootstrap
 import io.dropwizard.setup.Environment
 import ru.vyarus.dropwizard.guice.GuiceBundle
-import ru.vyarus.dropwizard.guice.module.installer.WebInstallersBundle
 import ru.vyarus.dropwizard.guice.support.TestConfiguration
 import ru.vyarus.dropwizard.guice.support.feature.DummyService
 import ru.vyarus.dropwizard.guice.support.util.BindModule
@@ -19,7 +18,7 @@ class ServletsApplication extends Application<TestConfiguration> {
     void initialize(Bootstrap<TestConfiguration> bootstrap) {
         bootstrap.addBundle(GuiceBundle.<TestConfiguration> builder()
                 .enableAutoConfig("ru.vyarus.dropwizard.guice.support.web.feature")
-                .bundles(new WebInstallersBundle())
+                .useWebInstallers()
                 .modules(new WebModule(), new BindModule(DummyService))
                 .build()
         );