+++ CHANGELOG.md
@@ -16,6 +16,7 @@
     - GuiceBundle noDefaultInstallers() shortcut method or GuiceyOptions.UseCoreInstallers option directly may be used to disable installation
 * Add methods to GuiceBundle builder:
     - printAvailableInstallers() - diagnostic reporting configured to show only available installers (to easily spot available features)
+    - useWebInstallers() - shortcut for installing WebInstallersBundle
     
     
 ### 3.3.0 (2016-08-02)